feat: basic implementation of entire schema to query flow
What the api can do at as of this commit:

- Admins can create/delete users
- Admins can create/delete/list orgs
- Admins can generate org api tokens
- Admins can add a schema and query to an org
- Orgs can upload data against an admin-added schema
- Orgs can run a admin-added query

Some background as to why there is so much in this commit. This represents output from a weekend hackathon squashed to avoid an unnecessarily noisy and chaotic commit history.

I started with Bun and Biome because they significantly simplified tooling and increased speed. However, Bun debugger support and test suite IDE integration is lacking. It became difficult to debug and so I reverted back to `vitest` and `tsx` which seems to be working well and still offering simplification and speed benefits. I have also used Hono as the web api. It is similar to express' api but much faster and more ergonomic. It is holding up well.