revert(cache): allow custom Ttl on cached items
cd4288cb and 6d449c33 migrated the cache to Temporal and also removed the ability for the caller to set a custom ttl. This caused the root account to disappear from the cache meaning, for a new node, if the admin wasnt' created in < 1 minute then it could not be created via the endpoint.