fix(soarchain): add optimised aggregations
Soarchain are running weekly queries and their dataset is growing quickly. Because we're mid fligth there is insufficient time to correct the schemas and we need fixes for type coercion to landed. So as a hotfix, this adds optimised queries to manage runtime:

1. Uses `collection.estimatedDocumentCount()` over costly aggregation pipeline
2. Uses Dates to limit the `Best Breaking Query` to the week in question

We will aim to do the same for the remainder of their competition via this method and then revert once the competition is complete and we can roll out a better solution.