test(nilcomm): add flag to enable nilcomm in a test suite, remove as test env feature
If nilcomm is required in a test it should be enabled via a test fixture parameter because enabling it via .env.test causes conflicting consumers and producers in multiple suites. This isn't an ideal solution and will need to be matured in time.

This also now correctly closes the fixtures mq channel and connection.